Move container/stream to daemon/internal/stream
