+++ container/exec.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/containerd/containerd/v2/pkg/cio"
 	"github.com/containerd/log"
-	"github.com/docker/docker/container/stream"
+	"github.com/docker/docker/daemon/internal/stream"
 	"github.com/docker/docker/libcontainerd/types"
 	"github.com/docker/docker/pkg/stringid"
 )
