+++ daemon/internal/stream/attach.go
